tutorials/cloud_deployment/aws_sagemaker/gat_cora/code/requirements.txt
pyTigerDriver==1.0.14
pyTigerGraph[gds]==0.9
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.13
